Port VideoToolbar components to TypeScript
Currently unused, not sure what to do with these
