+++ src/components/Video/VideoSizeButton.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { useTranslator } from '@u-wave/react-translate';
 import Tooltip from '@mui/material/Tooltip';
 import IconButton from '@mui/material/IconButton';
